# specs/001-core-framework-homepage/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: API Architect - Core Framework API
  description: |
    REST API for the API Architect Core Framework & Homepage feature.
    Provides endpoints for health checks, module configuration, and authentication shell.
  version: 0.1.0
  contact:
    name: API Architect Team
  license:
    name: MIT
    identifier: MIT
servers:
  - url: http://localhost:8000
    description: Local development server
  - url: https://api.api-architect.dev
    description: Development environment
...
paths:
...
  /api/v1/modules:
    get:
      operationId: listModules
      summary: List all modules
      description: Returns all registered application modules with their configuration
      tags:
        - Modules
      security:
        - bearerAuth: []
      parameters:
        - name: enabled
          in: query
          description: Filter by enabled status
          required: false
          schema:
            type: boolean
      responses:
        '200':
          description: List of modules
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ModuleListResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/v1/modules/{moduleId}:
    get:
      operationId: getModule
      summary: Get module by ID
      description: Returns a specific module's configuration
      tags:
        - Modules
      security:
        - bearerAuth: []
      parameters:
        - name: moduleId
          in: path
          description: Module identifier
          required: true
          schema:
            type: string
            pattern: '^[a-z][a-z0-9-]*$'
            example: requirements-studio
      responses:
        '200':
          description: Module details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ModuleResponse'
        '401':
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          description: Module not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT bearer token authentication.
        For development, use mock token: `mock-token-12345`
  schemas:
...
    ApiMeta:
      type: object
      description: Response metadata
      required:
        - requestId
        - timestamp
      properties:
        requestId:
          type: string
          format: uuid
          description: Request correlation ID
          example: '550e8400-e29b-41d4-a716-446655440000'
        timestamp:
          type: string
          format: date-time
          description: Response timestamp
          example: '2025-12-10T00:00:00Z'
    ApiError:
      type: object
      description: Error information
      required:
        - code
        - message
      properties:
        code:
          type: string
          description: Error code
          example: UNAUTHORIZED
        message:
          type: string
          description: Human-readable error message
          example: Authentication required
        details:
          type: object
          description: Additional error context
          additionalProperties: true
    ErrorResponse:
      type: object
      description: Standard error response
      required:
        - error
        - meta
      properties:
        error:
          $ref: '#/components/schemas/ApiError'
        meta:
          $ref: '#/components/schemas/ApiMeta'

    # Health schemas
...
    Module:
      type: object
      description: Application module
      required:
        - id
        - name
        - description
        - icon
        - route
        - enabled
        - order
      properties:
        id:
          type: string
          description: Unique identifier (kebab-case)
          pattern: '^[a-z][a-z0-9-]*$'
          minLength: 1
          maxLength: 50
          example: requirements-studio
        name:
          type: string
          description: Display name
          minLength: 1
          maxLength: 100
          example: Requirements Studio
        description:
          type: string
          description: Short description
          maxLength: 200
          example: Define and manage API requirements in plain text
        icon:
          type: string
          description: Lucide icon name
          minLength: 1
          maxLength: 50
          example: FileText
        route:
          type: string
          description: Navigation route path
          pattern: '^/app/[a-z][a-z0-9-/]*$'
          example: /app/requirements
        enabled:
          type: boolean
          description: Whether module is accessible
          example: true
        order:
          type: integer
          description: Display order
          minimum: 1
          maximum: 100
          example: 1
        badge:
          type: string
          description: Optional badge text
          maxLength: 20
          example: New
    ModuleResponse:
      type: object
      required:
        - data
        - meta
      properties:
        data:
          $ref: '#/components/schemas/Module'
        meta:
          $ref: '#/components/schemas/ApiMeta'
    ModuleListResponse:
      type: object
      required:
        - data
        - meta
      properties:
        data:
          type: array
          items:
            $ref: '#/components/schemas/Module'
        meta:
          $ref: '#/components/schemas/ApiMeta'

    # Auth schemas
...